# internal/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Dumpster API
  description: Multimodal knowledge base REST API
  version: "1.0.0"
servers:
  - url: /
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >-
        An anonymous session cookie minted server-side on the first request.
        No login or registration is required: the API automatically issues a
        new session_id cookie whenever the request arrives without one or
        presents an expired session. The cookie drives RLS isolation so each
        session only sees its own knowledge bases, documents, and chunks.
  schemas:
...
    IntrusionTestCommunity:
      type: object
      description: >-
        One tested community's word/topic intrusion test outcome: a handful
        of its real members were shown to a judge alongside one entity
        deliberately taken from a different community (the intruder), and
        the judge was asked to spot which one doesn't belong.
      properties:
        community_id:
          type: integer
        members:
          type: array
          items:
            type: string
        intruder_text:
          type: string
        judge_answer:
          type: string
        correct:
          type: boolean
      required: [community_id, members, intruder_text, judge_answer, correct]
    IntrusionTestResult:
      type: object
      description: >-
        A knowledge base's most recent intrusion test run -- a quantitative
        coherence metric for its detected communities. computed_at is null
        and communities is empty when a test has never been run, or when
        fewer than two communities were large enough to test.
      properties:
        computed_at:
          type: string
          format: date-time
          nullable: true
        score:
          type: number
          description: Correct / tested_count, in [0, 1]. Zero when tested_count is zero.
        tested_count:
          type: integer
        communities:
          type: array
          items:
            $ref: "#/components/schemas/IntrusionTestCommunity"
      required: [computed_at, score, tested_count, communities]
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
paths:
...
  /kbs/{id}/intrusion-test:
    get:
      summary: Get a knowledge base's most recently run intrusion test result
      description: >-
        Returns a zero-value result (null computed_at, empty communities)
        rather than a 404 when an intrusion test has never been run for
        this knowledge base yet — that's a normal first-run state, not an
        error.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The knowledge base's most recent intrusion test result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrusionTestResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Knowledge base not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: >-
        Run a word/topic intrusion test over a knowledge base's current
        community structure
      description: >-
        Manually triggered: for each of up to 30 communities large enough
        to test, mixes a few of its real top members with one entity taken
        from a different community and asks an LLM judge to spot the
        intruder. The fraction of communities where the judge succeeds is a
        quantitative coherence metric for the current community structure,
        meant for comparing before/after a pipeline change (chunking, edge
        weighting, entity filtering) rather than eyeballing examples.
        Sampling and intruder selection are deterministic for a given
        community structure, so re-running this against unchanged data
        always produces the same score. Fully replaces whatever result was
        there before. Requires community detection to have already been run
        at least once for this knowledge base.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The newly computed intrusion test result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrusionTestResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Knowledge base not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: >-
            Community detection has never been run for this knowledge base,
            so there is no community structure to test.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...